Added speed limit of 30m/s to Spawnpoint Scanning.
This commit adds a method `assign_spawns' to assign a worker id to each
spawn point. The goal is to keep each worker's scanning speed to be
below a set limit (defaults to 30m/s). The algorithm greedily tries to
assign each spawnpoint to the account that need to travel the slowest to
scan that location. If this cannot be done, an additional delay will be
added. User will be prompted to increase the number of account used if
such delay is larger than 60s.